fix: move ILatestResponse to a separate folder
